feat(maps-service): move to a google-maps api service
